move filter button out into own component
